Moved the PlayerBullet class
